requirements.txt
fastapi
uvicorn
requests
streamlit
pydantic
python-dotenv
langchain
langgraph
langchain_groq
langchain_community